cli/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "lego-image-processor"
version = "0.1.0"
description = "A CLI tool that converts images to LEGO-compatible color palettes using advanced color quantization algorithms"
authors = ["LEGO Image Processor Team"]
readme = "README.md"
packages = [{include = "lego_image_processor", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
Pillow = "^10.0.0"
numpy = "^1.24.0"
scikit-image = "^0.21.0"
click = "^8.1.0"
tqdm = "^4.65.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-benchmark = "^4.0.0"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
lego-image-processor = "lego_image_processor.cli.main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/lego_image_processor"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
]
